remove path tools
